refactor: move store to heap module
